[Task] Use Image URI generated by ImageService for non-processed files
If there are non-processed files (e.g. gifs) recently the URI of the image was absolute. This works on systems, where the filedamin can be found in the root. If that's not the case, the images cannot be loaded in the frontend.

This change fixes this behaviour, so the absolute path is used.

(cherry picked from commit 0f83dfe9c3a86c5811905163fb370690a0e9601f)